# Add noasm build tag