--- a/asm/daxpy.go
+++ b/asm/daxpy.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build !amd64
+//+build !amd64 noasm
 
 package asm