--- a/asm/daxpy_amd64.go
+++ b/asm/daxpy_amd64.go
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+//+build !noasm
+
 package asm
 
 func DaxpyUnitary(alpha float64, x, y []float64)