This allows easy benchmarking against the pure pure Go implementation and also provides a way to easily assess compiler improvements.

e.g.

	$ cd $GOPATH/src/github.com/gonum/blas/native
	$ go test -bench . > asm.txt
	$ go test -tags noasm -bench . > noasm.txt
	$ benchcmp noasm.txt asm.txt