@kortschak 2015-01-11
PTAL @fhs 

Please rebase this onto #3 when this is merged.